smaller file sizes refactor, tweaks for SSA, reduced comments
