+++ ProcessSSA.py
@@ -1,84 +1,12 @@
 import re
-import datetime
-import json
-import warnings
+import Subtitling
 
-#a standard class to act as an intermediary between formats
-class SubLine:
-  def __init__(self, subs, start, end):
-    self.dialogue = []
-    for x in subs:
-      self.dialogue.append(x)
-    self.time = {
-      'start': start,
-      'end': end
-    }
-  def hms_to_frames(self, framerate):
-    return framerate
-  def hms_to_seconds(self):
-    return 0
-  def __repr__(self):
-    out = "[SUBLINE]=>{" +\
-      " start:" + str(self.time['start']) + \
-      ", end:" + str(self.time['end']) + \
-      ", dialogue:" + str(self.dialogue) + \
-      "}\n"
-    return out
+''' Inputs (converts to an array of SubLine class files) '''
 
-def read_full_file(sub_file):
-  f = open(sub_file, 'r')
-  text = f.read()
-  f.close()
-  return text
-def write_file(sub_file, sub_text):
-  f = open(sub_file, 'w')
-  f.write(subtext)
-  f.close()
-  return subtext
-
-#type unspecific, returns datetime object
-def subs_to_time(subs_time):
-  hours = int(subs_time[0])
-  minutes = int(subs_time[1])
-  seconds = int(subs_time[2])
-  microsec = int(subs_time[3])*1000#srt is in thousandths of a second
-  return datetime.time(hours, minutes, seconds, microsec)
-
-#reads file, returns dictionary
-def read_json(json_file):
-  return json.loads(read_full_file(json_file))
-
-'''
-  READING SRT
-'''
-# <<TODO>> SRT is a fairly simplistic format but there's definitely room for validation and error handling here
-def process_srt(sub_text):
-  #a blank line represents a next section
-  basic_split = sub_text.strip().split('\n\n')
-  formatted_lines = []
-  for subtitle in basic_split:
-    subtitle = subtitle.split('\n')
-    times_str = subtitle[1].split('-->')
-    time_formatted = []
-    for time_str in times_str:
-      time_str = time_str.strip().replace(',', ':').split(':')
-      time_formatted.append(subs_to_time(time_str))
-    dialogue_lines = []
-    for dialogue_index in range(2, len(subtitle)):
-      dialogue_lines.append(subtitle[dialogue_index])
-    formatted_lines.append(SubLine(dialogue_lines, time_formatted[0], time_formatted[1]))
-  return formatted_lines
-
-'''
-  READING SSA
-'''
-# formatting from here https://matroska.org/technical/specs/subtitles/ssa.html
-# kinda hacky breakdown to find the individual parts, may require an overhaul
 def process_ssa(sub_text):
   formatted_lines = []
-  #a blank line represents a next section
   segment_split = re.split('\n\[Events\]\n|\n\[V4 Styles\]\n', sub_text)
-  dialogue_lines=segment_split[2].split('\n') # <<TODO>> Remove hardcoding for flexibility
+  dialogue_lines=segment_split[2].split('\n')
   dialogue_format=parse_ssa_line(dialogue_lines.pop(0))
   #get indexes of relevant format pieces
   for subtitle in dialogue_lines:
@@ -89,11 +17,9 @@ def process_ssa(sub_text):
     start_time = format_ssa_time(subtitle['Start'])
     end_time = format_ssa_time(subtitle['End'])
     dialogue = [subtitle['Text']]
-    formatted_lines.append(SubLine(dialogue, start_time, end_time))
+    formatted_lines.append(Subtitling.SubLine(dialogue, start_time, end_time))
   return formatted_lines
 
-#formats an SSA line from a string into its parts, dialogue additional commas handled
-# <<TODO>> Error handling for poorly specified files, return an incomplete result?
 def parse_ssa_line(line, formatting=None):
   parsed_line = {
     'title': "",
@@ -112,42 +38,14 @@ def parse_ssa_line(line, formatting=None):
       parsed_line['values'].append(entry.strip())
   return parsed_line
 
-#formats to milliseconds to match srt specs
 def format_ssa_time(time_str):
   fmt_time = time_str.replace('.', ':').split(':')
   fmt_time[3]+= "0"
-  return subs_to_time(fmt_time)
+  return Subtitling.subs_to_time(fmt_time)
 
 
-###---OUTPUT FORMATTING--###
-#regular converters, reformats subtitles to a string in the correct format
+''' Outputs '''
 
-'''
-  WRITING SRT
-'''
-def format_srt(sub_lines):
-  srt_output = ""
-  counter = 1
-  for entry in sub_lines:
-    if counter != 1:
-      srt_output += "\n"
-    srt_output += str(counter)+"\n"
-    srt_output += entry.time['start'].strftime("%H:%M:%S,%f")[0:12]#hacky solution to switch from microseconds to milliseconds
-    srt_output += " --> "
-    srt_output += entry.time['end'].strftime("%H:%M:%S,%f")[0:12]
-    srt_output += "\n"
-    for line in entry.dialogue:
-      srt_output += line
-      srt_output += "\n"
-    counter += 1
-  return srt_output
-
-
-'''
-  WRITING SSA: A long convoluted section
-'''
-
-#tests need to be updated when this is expanded upon
 def format_ssa(sub_lines):
   ssa_skeleton = get_ssa_format()
   ssa_output = ssa_skeleton["info_string"] + "\n\n"
@@ -164,7 +62,6 @@ def format_ssa(sub_lines):
       if(len(entry.dialogue)>1):
         ssa_output += "\\n"
     ssa_output+="\n"
-  print ssa_output
   return ssa_output
 
 #returns skeletal strings built from config specification to apply time and dialogue to for SSA
@@ -245,10 +142,11 @@ def get_ssa_format():
     "events": ssa_format_event(event)
   }
 
-#<<TODO>> style and event violate DRY really
-#<<TODO>> very hacky substrings
 def ssa_format_title(info):
-  return "[Script Info]\nTitle: "+info['title']+"\nOriginal Script: "+info['Original Script']+"\nScriptType: v4.00"
+  return "[Script Info]\nTitle: " \
+        + info['title']+"\nOriginal Script: " \
+        + info['Original Script'] \
+        + "\nScriptType: v4.00"
 def ssa_format_style(style):
   format_string = "Format: "
   style_string = "Style: "
@@ -266,8 +164,5 @@ def ssa_format_event(event):
   return {"format": "[Events]\n"+format_string[:-2], "event_shell": event_string}
 
 
-'''
-  WRITING SUB: Pretty simple with a time to frame conversion
-'''
