+++ test_files/INFO.md
@@ -1,0 +1,15 @@
+
+### Initial Files
+- Merge_Match.srt
+  Base file for merges
+- Merge_1sec.srt
+  One second off from Marge_Match, for testing pairing functions
+- Merge_LongLine.srt
+  Very long line, for testing character limits
+- Merge_Multiline.srt
+  For testing means of handling multiple line conversion
+
+### Result Files
+- Result_Merge_Match.srt
+  The result of merging Merge_Match.srt with itself or Merge_1sec.srt if a one second window is allowed
+
