+++ test_files/Result_Merge_Match.srt
@@ -1,0 +1,14 @@
+1
+00:01:12,833 --> 00:01:19,000
+Los abrazos rotos
+Los abrazos rotos
+
+2
+00:01:42,458 --> 00:01:49,417
+(Paginas del periodico)
+(Paginas del periodico)
+
+3
+00:01:54,708 --> 00:01:57,750
+¿Cómo te llamas?
+¿Cómo te llamas?
